+++ tethysdash_plugin_geoglows/data/map_configs.json
