+++ tethysdash_plugin_geoglows/utils/plots.py
@@ -5,10 +5,8 @@ import pandas as pd
 import numpy as np
 import scipy.stats as stats
 import plotly.graph_objects as go
-import json
 import getpass
 import pwd
-from pyproj import Transformer
 
 
 username = os.environ.get("NGINX_USER", getpass.getuser())
@@ -468,44 +466,3 @@ def plot_ssi_one_month_each_year(reach_id, month):
         yaxis_title='SSI'
     )
     return fig
-
-
-def load_country_list():
-    filename = os.path.join(module_path, "countries_extents.json")
-    country_list = []
-    with open(filename, "r") as file:
-        data = json.load(file)
-    for country_name, _extent in data.items():
-        country_list.append({"value": country_name, "label": country_name})
-    return country_list
-
-
-def load_country_extents():
-    filename = os.path.join(module_path, "countries_extents.json")
-    country_extent = {}
-    with open(filename, "r") as file:
-        data = json.load(file)
-    for country_name, extent in data.items():
-        if extent:
-            extent[0] -= 0.1
-            extent[1] -= 0.1
-            extent[2] += 0.1
-            extent[3] += 0.1
-        country_extent[country_name] = extent
-    return country_extent
-
-
-def convert_4326_to_3857(lon, lat):
-    """
-    Convert WGS84 (EPSG:4326) to Web Mercator (EPSG:3857).
-
-    Parameters:
-        lat (float): Latitude in degrees.
-        lon (float): Longitude in degrees.
-
-    Returns:
-        (x, y): Tuple of coordinates in meters.
-    """
-    transformer = Transformer.from_crs(4326, 3857, always_xy=True)
-    x, y = transformer.transform(lon, lat)
-    return x, y
